Fix a possible bit of undefined behavior

+++ include/num.h
@@ -138,7 +138,7 @@ typedef struct BcNum {
 
 #define BC_NUM_KARATSUBA_ALLOCS (6)
 
-#define BC_NUM_CMP_SIGNAL (~SSIZE_MAX)
+#define BC_NUM_CMP_SIGNAL (~((size_t) SSIZE_MAX))
 
 #define BC_NUM_ROUND_POW(s) (bc_vm_growSize((s), BC_BASE_POWER - 1))
 #define BC_NUM_RDX(s) (BC_NUM_ROUND_POW(s) / BC_BASE_POWER)
